Bug #14062726 FAILING ASSERTION: TRX_STATE_EQ(TRX, TRX_STATE_ACTIVE)
New asserts were added for XA processing in InnoDB for 5.6 which revealed
this bug in the server.  The RELEASE SAVEPOINT command needs to have
the same checks for XA state as ROLLBACK TO SAVEPOINT so that
a savepoint is not released while the transaction is in the prepared state. 

In addition, both ROLLBACK TO SAVEPOINT and RELEASE SAVEPOINT should
be allowed while the XA transaction is ACTIVE.