Bug#14250296 - ASSERT IN GET_DIGEST_TEXT DURING CONCURRENT DML WORK LOAD
Before this fix, the code displaying a digest text could fail with an assert,
on race conditions.
An example of race condition is:
- thread 1 executes a statement A then statement B
- thread 2 performs a SELECT * from events_statements_current,
which displays the last statement executed by thread 1, 
including the statement digest.

When thread 2 reads the digest data at the very same moment statement B starts,
thread 1 could read a NULL charset, which triggered the assert.

The digest structure used is not locked, and this is expected.
The proper fix is to sanitize every data upon read, to make sure it is valid,
as underlying buffers are expected to change concurrently.
This fix changes the way CHARSET_INFO is used:
- a charset number instead of a pointer is stored,
- the digest charset is sanitized on read.