BUG#14343457: SELECT @@SESSION.GTID_DONE ASSERT FAILURE WHEN EXECUTED ON ONGOING TRANSACTION
SELECT @@SESSION.GTID_DONE executed within an on-going transaction was
causing assertion failures because the on-going transaction did not have
a 'valid' GTID assigned to it. The GTID is only assigned upon commit
when the transaction is written to the binlog.

Notice, however, that the assert failure does not happen if a
SET @@SESSION.GTID_NEXT is executed as this operation assigns a 'valid'
GTID to the next transaction.

Fixed Group_cache::get_gtids(Gtid_set) method to consider the case on
which group has not yet GTID assigned.