BUG#14298750: SLAVE_MASTER_INFO NOT UPDATED WHEN SLAVE ATTACHED TO MASTER
Problem: When setting up replication on a master
and slave where the slave is using
master-info-repository=TABLE, the table is not
updated when replication is started. This is because
the table mysql.slave_master_info is not synchronized
automatically when @@GLOBAL.sync_master_info=0 and this
causes applications reading this table to fail or behave
incorrectly.     

Solution: We fix this problem by synchronizing
mysql.slave_master_info during CHANGE MASTER command.