BUG#14364558 - ASSERT `TABLE_LIST->PRELOCKING_PLACEHOLDER == FALSE'                FAILED IN CHECK_LOCK_AND_START_STMT
Problem:
rqg_signal_resignal test hits the assert 
DBUG_ASSERT(table_list->prelocking_placeholder == false), in
check_lock_and_start_stmt();

Analysis:
The above assert was added as part of fix for Bug#13036505,
to capture the fact that prelocking_placeholder should always
be false in check_lock_and_start_stmt() in theory. But in practice,
this assertion is hit. Even without the fix Bug#13036505 this
problem exists. So, this problem is not an effect to Bug#13036505,
and hence can be treated as a new issue.

For now we shall remove this assert which will keep the
old behavior. And we keep this bug open, so as to
dig into the issue more and provide proper fix.

Current observation is that, the problem is not reproducible with
--threads=[<=3] and could be due to concurrent operations.

perl runall.pl --threads=10 --queries=1M --duration=300 
--grammar=conf/runtime/signal_resignal.yy 
--mysqld=--max-sp-recursion-depth=10 
--reporters=Deadlock,ErrorLog,Backtrace,Shutdown 
--basedir=$BASEDIR --vardir=vardir  --mysqld=--log-output=file 
--mysqld=--loose-lock-wait-timeout=1 
--mysqld=--loose-innodb-lock-wait-timeout=1 
--testname=rqg_signal_resignal