Bug#11753308: MYSQL_KILL() AND COM_PROCESS_KILL ONLY HANDLE 32-BIT IDS
thread_id / CONNECTION_ID() is currently 32 bit on the wire, but may be 64 bit in the API,
depending on the definition of ulong on a given platform. If we silently drop half the
bits on the floor, mysql_kill() may ask for an incorrect (and usually long expired) thread
ID to be terminated.

We now guard against some 32/64 issues.
We do this in both the client and the server to address mixed scenarios of old server /
new client and vice versa.

Using SQL instead of the C API (SELECT CONNECTION_ID() / KILL CONNECTION ...) should
produce better results. mysqladmin already does this.