Bug#11745768: WRONG DATATYPE RETURNED FOR CONNECTION_ID()?
SQL-function CONNECTION_ID() assumed that its value would always be 32-bit,
while on some systems (where ulong is 64-bit), it's really not.

This patch corrects both the field-size, and the signedness.

The test shows the new data-format; the code was also tested with an
actual 64-bit value for thread_id in the debugger.