BUG#11749055 - INSERT SELECT ON DUPLICATE KEY UPDATE, LOAD DATA REPLACE                WRONG LOCK
BACKGROUND:
INSERT SELECT ON DUPLICATE KEY UPDATE and LOAD DATA CONCURRENT REPLACE
uses TL_WRITE_CONCURRENT_INSERT lock though it either update/inserts
or deletes/inserts a row respectively.

ANALYSIS:
When a SELECT SUM(column_name) UNION SELECT SUM(column_name)
query is issued and while it is running if following two commands 
are run 
1. INSERT SELECT ON DUPLICATE KEY UPDATE
2. LOAD DATA CONCURRENT REPLACE
Second SELECT SUM(column_name) gets different values from the 
first SELECT query as the tuples in the table have either been 
updated/inserted OR deleted/inserted in the table repectively,
when the above SELECT..UNION query is running
Same issue exists with simple INSERT.. ON DUPLICATE KEY UPDATE
when it is used in stored function or trigger.

This behaviour is not desired i.e Both SELECT's in this union 
query should have the same view of the table when it starts.

The root cause for this bug INSERT or INSERT SELECT with ON 
DUPLICATE KEY UPDATE clause and LOAD DATA CONCURRENT REPLACE 
should not be doing updates/deletes respectively with 
TL_WRITE_CONCURRENT_INSERT lock.

FIX:
This bug is fixed by changing the default lock type of INSERT or 
INSERT..SELECT statement to TL_WRITE_DEFAULT lock in the parser
itself when they are issued with ON DUPLICATE KEY UPDATE clause.
Same is taken care for LOAD DATA CONCURRENT REPLACE statement 
as well.
NOTE:
When CONCURRENT is specified with REPLACE in LOAD DATA statement
CONCURRENT is overridden by REPLACE.