Bug#14078793 CREATE_TMP_TABLE() ADDS MORE COPY_FUNCS THAN IT ALLOCATES SPACE FOR
Problem: In some cases, create_tmp_table() is called with incorrect
parameters, causing it to overflow the memory allocated for copy_func
and overwrite other variables.

The number of fields and functions to copy is counted by
count_field_types() and stored in a TMP_TABLE_PARAM object. In
create_tmp_table(), these numbers are used to allocate memory for
storing pointers to the fields and function items. When items are
changed in a way that changes the type, the count must be
updated. When it isn't, create_tmp_table() may write more items into
copy_func than memory has been allocated for.

Fix: Add assertion to check that copy_func never overflows and call
count_field_types() to get a recount when items may have changed.

No new tests are added since this bug is covered by existing tests.