Bug#14392805 ASSERT IN ROLLBACK_INPLACE_ALTER_TABLE
When an error (such as a duplicate key error) is detected while
applying the modification log during online index creation, we would
clear the 'to_be_dropped' flags on any indexes that were to be dropped
in the same ALTER TABLE statement. This would cause a debug assertion
failure during rollback. The fix is to remove the debug assertion on
rollback.

rb:1208 approved by Sunny Bains