Bug #14219233 FOREIGN KEY...CASCADE OR SET NULL MAY CONFLICT WITH ONLINE DDL
For "ALTER .... INPLACE (LOCK=NONE/DEFAULT) we avoid online inplace and 
so force it to use SHARED lock.

This is done to protect ALTER TABLE INPLACE against a bug in
MySQL MDL that fails to lock foreign key table while modifying
the parent table in a way that affects foreign table.
(For example: delete from parent; will affect child table if
there is fk-del-on-cascade/set-null dependency on child table.)
If avoided, this leads to an issue in form that parallel
transaction running ALTER TABLE is allowed to proceed on TABLE
that is already being modified by other transaction, further 
causing data-inconsistency.

Remove this fix once WL#6049 is fixed as it would take
care of the issue in MySQL MDL locking.

This bug also update TC for bug#14189862 as they are related.