Bug#14272788: Query with MaterializeScan and materialized subquery               returns too many rows
Patch #8 - Fix problem with how subqueries are transformed to semi-join.

There was a problem in how JOIN::flatten_subqueries() determined whether
to transform a subquery to semi-join or not. When checking number of
tables in query block after a transformation, the tables already added
were not considered.

This is now changed so that the first loop decides whether a subquery
qualifies for transformation or not, and the second and third loops
reuse the decision from the first loop. In addition, if a join nest
with fewer tables follows one with more tables, which was not transformed,
the smaller nest is now considered. In addition, the condition
"< MAX_TABLES" was checked, but it should be "<= MAX_TABLES".

sql/item_subselect.h
  Added the sj_chosen member variable that tells whether a subquery is
  chosen for semi-join transformation.

sql/sql_optimizer.cc
  Changes to JOIN::flatten_subqueries(), see above.