Bug #13450566-UPDATING 16M BLOB CAUSES TOO MANY RECURSIVE X-LOCKS
This changes the way rw-locks keep track of recursive write or
x-locks by decrementing recursive x-locks by 1 instead of
X_LOCK_DECR (0x00100000).  But this decrement by 1 only happens
with the third x-lock by the same thread, which is the second
recursive x-lock.

Instead of subtracting X_LOCK_DECR for every recursive lock, this
patch subtracts it once for the initial x-lock, once for the first
recursive x-lock, and by just 1 thereafter.  So now the ranges for
lock_word are like this this;

X_LOCK_DECR   = not locked
1 to X_LOCK_DECR-1
              = s-locked by X_LOCK_DECR-n threads
0             = x-locked once
-X_LOCK_DECR+1 to -1
              = reserved by an x-locking thread but still has s-locks
-(2*X_LOCK_DECR) to -X_LOCK_DECR
              = recursively x-locked by a single thread

Only the last range listed above has changed.  The rest are the same
as before.