Fix for bug #14404635  "ASSERTION `!MAYBE_NULL == !KEY_INFO[*ADD]. KEY_PART[KEY_PART]. FIELD->REAL_MAYBE_".
Attempt to add an index on InnoDB table and change nullability of
column participating in this index in the same ALTER TABLE led to
assertion failure.

The assert was firing because code in innobase_create_key_defs()
assumed that information about nullability of column in
KEY_PART_INFO object for a column in new index is consistent
with a state of Field object for the same column referenced by
KEY_PART_INFO::field member. This was not true in general case
since for columns which existed in table before ALTER
KEY_PART_INFO::field member pointed to Field object respresenting
old, unchanged version of column. While other information in
KEY_PART_INFO corresponds to new version of column, with changes
taken into account.

This patch fixes the problem by changing code to always store in
KEY_PART_INFO::field pointers to Field object which correspond to
new version of columns.
Unfortunately, InnoDB's code responsible for reporting duplicate
key errors during in-place ALTER TABLE relied on old behavior, i.e.
that KEY_PART_INFO::field contained Field object which was bound to
TABLE::record[] for old version of table.
So this patch also changes InnoDB code for in-place ALTER TABLE to
use record buffer for new version of table for duplicate key
reporting.