Bug#13984150 - PERFSCHEMA.PFS_UPGRADE TEST FAILS WITH "UPGRADE FAILED" MAINLY IN GCOV
Before this fix, the test pfs_upgrade would test different upgrade cases
involving:
- a table
- a view
- a stored procedure
- a stored function
- an event
in the same script, which resulted in 5 execution of mysql_upgrade.

The problem was that, when doing special builds involving time consuming
instrumentation (valgrind builds, gcov builds), and given that mysql_upgrade
is an long operation, the test script did time out and fail.

The fix is to split this long test in smaller peices,
to avoid time outs and improve manageability.

This fix is a test script fix only, no server code was changed.