Bug#14501403 CRASH IN ITEM_CACHE_STR::SAVE_IN_FIELD WITH NATURAL JOINS, SET, VARBINARY
Regression introduced by fix for bug #13013970.

Problem: Item_cache_str::save_in_field() dereferences a null pointer
if the cached value is NULL.

When the cached value is NULL, null_value is set to true and value is
set to a null pointer. Item_cache_str::save_in_field() tries to find
the length of the String value points to without checking for NULLs
first. This leads to dereferencing a null pointer.

Fix: Check that the value pointer is not null before trying to find
the length of the String it points to.