Bug #14458217:  password validation plugin: crash when with password(null)
The code in PASSWORD() and OLD_PASSWORD() was passing
a NULL pointer to the password validation plugin(s) instead 
of a String class instance when a NULL is passed as an 
argument and this was not expected by the plugin.

Made sure all calls to check_password_policy() (in both
PASSWORD() and OLD_PASSWORD()) always get an non-null
String class by passing empty string when a NULL argument
is supplied to these functions.
Added debug assertions that all password validation plugin
API calls always get non-null Strings.
Test case added.
Updated the test suite plugin list to allow platform independent
loading of the password validation plugin.