Fix for Bug#14503142 PERFORMANCE REGRESSION FOR DESCENDING RANGE SCAN WHEN                      USING ICP
The performance regression is caused by use of ICP in combination with
a descending range scan that is using QUICK_SELECT_DESC. The
QUICK_SELECT_DESC object executes the range scan by reading the index
using ha_index_prev() and then evaluating if the record is within the
current range. This works fine when we have not pushed any index
condition to the storage engine since every record will be returned
and evaluated by QUICK_SELECT_DESC.get_next(). When ICP is used, the
call to ha_index_prev() will only return records that do not get
filtered by the pushed index condition. In this case, as soon as
QUICK_SELECT_DESC.get_next() tries to read the first record that is
outside the current range, the ICP implementation in the storage
engine will filter it away and continue reading from the index without
finding any qualifying records. It will continue reading from the
index until it has read to the end of the index. The increased
execution time is caused by the ICP implementation reading to the
end of the index instead of stopping when it has reached the end of
the current range.

This patch solves this problem by adding a new handler call
handler::set_end_range() which informs the storage engine about the
end of the current index scan. The ICP implementation in the storage
engine will use this information and check that the next record is
within this range. For each range that QUICK_SELECT_DESC.get_next()
needs to read, it will use this call to let the storage engine know
about the end of the range.