This commit is a followup to revno 4960 / revid:jan.wedvik@oracle.com-20120822104942-q95h5zi9m729lkkk which was a fix for 'Bug#14190114: CLUSTER CRASH DUE TO NDBREQUIRE IN ./LOCALPROXY.HPP DBLQH (LINE: 234)'.
When running ndbautottest, the regression test for bug 14190114 
(runDropTakeoverTest()) would always fail. The reason for this was that
another regression test (runBug13416603()) would leave some data nodes
in a state where they would not start again after stopping because
of an error insert. This commit fixes this problem by adding cleanup
code to runBug13416603().