Bug#14500792 ASSERTION FAILED: (*REF)->IS_TEMPORAL(), FILE ITEM.CC, LINE 7581
Problem: A query with GROUP BY WITH ROLLUP comparing a grouping column
using the IN operator causes an assertion to fire.

During rollup processing, Item_null_result items are generated for the
NULL values in the rollup rows by JOIN::rollup_make_fields(). When
comparing using the IN operator, the grouping value is represented as
an Item_ref pointing to an Item_null_result.
Item_ref::val_time_temporal() asserts that its ref is a temporal item
by checking (*ref)->is_temporal(). This causes the field type to be
checked, and is_temporal_type() defines that MYSQL_TYPE_NULL is not a
temporal type. Therefore, the assertion fires.

Fix: Add a constructor to Item_null_result that sets the field and
result types so that they can be set to the correct type for that
column when instantiating Item_null_result in JOIN::rollup_init() and
JOIN::rollup_make_fields().