BUG#11758505: MISSING RELAY LOG FROM INDEX FILE MESSAGE ON STARTUP NOT EXPLICIT ENOUGH
Problem: The error message thrown when a relay-log file
is missing from the relay_log index file, is not explicit
enough to inform the user that the relay-log file is not
found in the index file.
Old error mesaage:
 "Could not find target log during relay log initialization"
      
Solution: We changed the error message to make sure the
correct information is conveyed to the user.

New error message:
"Could not find target log file mentioned in relay log info
in the index file '<index_file_name>' during relay log
initialization."