BUG#14116526: RPL.RPL_GTID_SQL_UNTIL_BEFORE_AFTER FAILS WITH RESULT LENGTH MISMATCH ON TRUNK.
There is a opportunity window between the moment on which current GTID
(thd->owned_gtid) is appended to logged_gtids and the commit to engines
is performed, what on slow/profiling platforms can make SELECT
SQL_THREAD_WAIT_AFTER_GTIDS(GTID) return before data reaches the 
database.

To close this opportunity window, when current GTID is appended to
logged_gtids it is not removed from owned_gtids, it is only removed
after commit. SQL_THREAD_WAIT_AFTER_GTIDS(GTID) returns when GTID is
logged (flushed to binary log) and not on owned_gtids (committed).

@ sql/binlog.cc
   Only after commit stage remove current GTID from owned_gtids set.
@ sql/rpl_gtid.h
   Added Owned_gtids::is_intersection(const Gtid_set *other) method
   signature and Gtid_iterator on Owned_gtids class.
   Renamed Gtid_state::update(THD *thd, bool commit) to
   Gtid_state::update_on_flush(THD *thd).
   Added Gtid_state::update_on_commit(THD *thd) and 
   Gtid_state::update_on_rollback(THD *thd) method signature.
@ sql/rpl_gtid_execution.cc
   Between binlog flush and commit stages logged_gtids and owned_gtids
   may intersect, so invariant's assert must be removed.
@ sql/rpl_gtid_owned.cc
   Added Owned_gtids::is_intersection(const Gtid_set *other) method.
@ sql/rpl_gtid_state.cc
   Gtid_state::update_on_flush(THD *thd) adds current GTID to logged
   gtids and notifies waiting threads.
   Gtid_state::update_on_commit(THD *thd) removes current GTID from
   owned_gtids and cleans thd->owned_gtid.
   Gtid_state::update_on_rollback(THD *thd) removes current GTID from
   owned_gtids, notifies waiting threads and cleans thd->owned_gtid.
@ sql/rpl_rli.cc
   On Relay_log_info::wait_for_gtid_set() also checks that wait set does
   not intersect with owned_gtids set.