This is a fix for Bug#14644936 "INEFFICIENT AND INCORRECT EXECUTION OF PUSHABLE OUTER JOINS".
This fix copies code from join_read_always_key(). This
new code checks if the key value that has been built for
access to the next table contains NULL values. If so, there
is no need to scan the next table as this would have retuned 
zero matching rows.

(There will be a regression test in a separate commit on a cluster branch,
as this bug can only be reproduced there.)