ndb - make online reorg change hashmap size if appropriate
Bug #14645319ONLINE REORGANIZE CAN NOT USE BIGGER HASHMAP ON OLD TABLES

before online reorg never changed hashmap size.

now it either keeps the old size or use the hardcoded
default hashmap size.

changing hashmap size only occur if the number of
fragments have increased and the old hashmap size
is not a multiple of the new fragment count.
also the bigger hashmap size must be a multiple of
the old hashmap size to guarantee that data are moved
from old fragments to new fragments only, therefore
the old hashmap size is used if that is not true.

this means that after an upgrade from a ndb version
supporting a smaller hashmap size to a ndb version
supporting a bigger hashmap size, will be downgradable
as long as no new tables are created or an online
reorg have been run after changing the number of
fragments (implictly by adding nodes, or changing maxrows,
or explicitly added partitions).

NOTE: neither unique index nor blob tables are reorganized.