Bug#14580971: CRASH IN SEL_ARG_RANGE_SEQ_NEXT() WHEN CALLED FROM               DSMRR_IMPL::DSMRR_INFO_CONST()
SEL_ARG_RANGE_SEQ::stack[] is used to store range predicates 
for up to MAX_REF_PARTS keyparts. Before this patch, 
sel_arg_range_seq_next() started to store in stack[1] which 
lead to overflow in the presence of MAX_REF_PARTS keyparts. 
This caused a crash.

The fix in this CS is to start storing in stack[0]. This is 
implemented by refactoring stack[] push and pop operations 
by hiding them behind an interface. This makes the code 
easier to understand.